requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
PyYAML==6.0.2
